requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
tensorflow==2.13.0
pillow==10.0.0
numpy==1.24.3



